feat: updated configuration and test
