+++ advanced_alchemy/__main__.py
@@ -1,9 +1,11 @@
-from advanced_alchemy.cli import alchemy_group
+from __future__ import annotations
+
+from advanced_alchemy.cli import add_migration_commands as build_cli_interface
 
 
 def run_cli() -> None:
     """Advanced Alchemy CLI"""
-    alchemy_group()
+    build_cli_interface()()
 
 
 if __name__ == "__main__":
